refactor(@embark/core): move console source out of lib folder
