+++ packages/core/console/tsconfig.json
@@ -7,7 +7,7 @@
   },
   "extends": "../../../tsconfig.base.json",
   "include": [
-    "src/lib/**/*"
+    "src/**/*"
   ],
   "references": [
     {
